Restructures cms app folders
