+++ gitpub/cms/views/courses.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.http import HttpResponse
 from gitpub.logging import debug
-from .models import Course
+from cms.models import Course
 
 # GET /courses
 @debug
